Add AWS batch executor

+++ run.nextflow.sh
@@ -2,7 +2,6 @@
 
 # DIR where the current script resides
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-
 batch_input=${1}
 pipeline=${2:-'nanopore'}
 profile=${3:-'codon'}
@@ -19,6 +18,17 @@ project_id=${10:-'prj-int-dev-covid19-nf-gls'}
 #################################
 echo ""
 echo "** Processing samples with ${DIR}/${pipeline}/${pipeline}.nf. **"
+
+if [ "$profile" = "awsbatch" ]; then
+      echo "** Retrieve secrets to ${DIR}/${project_id}-sa-credential.json **"
+      aws secretsmanager get-secret-value --secret-id $GOOGLE_APPLICATION_CREDENTIALS_SECRET_ID --query SecretString --output text > "$DIR/${project_id}-sa-credential.json"
+      gcloud auth activate-service-account --key-file="$DIR/${project_id}-sa-credential.json"
+      gcloud config set project ${project_id}
+      project_bucket="prj-int-dev-ait-eosc-aws-eval"
+      aws s3 cp ${batch_input} ${DIR}/data/ # download sample index file from s3 to local dir
+      aws s3 cp "s3://${project_bucket}/${dataset_name}/" "${DIR}/data/" --recursive --exclude "*/*"  # download projects_accounts and .fa files
+      batch_input="${DIR}/data/$(basename -- "$batch_input")" #local path to sample index file
+fi
 
 pipeline_dir="${root_dir}/${snapshot_date}/${pipeline}_${batch_index}"
 echo "** pipeline_dir: ${pipeline_dir} **"
@@ -37,13 +47,22 @@ nextflow -C "${DIR}/nextflow-lib/nextflow.config" run "${DIR}/${pipeline}/${pipe
 ########################################################################################
 # Update submission receipt and submission metadata [as well as all the analyses archived]
 ########################################################################################
-#"${DIR}/update.receipt.sh" "${batch_index}" "${snapshot_date}" "${pipeline}" "${profile}" "${root_dir}" "${dataset_name}" "${project_id}"
-#"${DIR}/set.archived.sh" "${dataset_name}" "${project_id}"
+"${DIR}/update.receipt.sh" "${batch_index}" "${snapshot_date}" "${pipeline}" "${profile}" "${root_dir}" "${dataset_name}" "${project_id}"
+"${DIR}/set.archived.sh" "${dataset_name}" "${project_id}"
 
-if [ "$profile" != "gls" ]; then
+if [[ "$profile" != "gls" && "awsbatch" ]]; then
       rm -R "${pipeline_dir}/workDir" &
       rm -R "${pipeline_dir}/storeDir" &
       rm -R "${pipeline_dir}/publishDir" &
       wait
       rm -R "${pipeline_dir}"
 fi
+
+if [ "$profile" = "awsbatch" ]; then
+      aws s3 rm --recursive "${pipeline_dir}/workDir" --quiet &
+      aws s3 rm --recursive "${pipeline_dir}/storeDir" --quiet &
+      aws s3 rm --recursive "${pipeline_dir}/publishDir" --quiet &
+      wait
+      aws s3 rm --recursive "${pipeline_dir}"
+fi
+
